requirements.txt
cffi==1.16.0
pycryptodome==3.19.0
cryptography==41.0.5
pycparser==2.21
